requirements.txt
absl-py==2.1.0
aiofiles==22.1.0
aiosqlite==0.18.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.9.0
cachetools==6.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.8
colorama==0.4.6
comm==0.1.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
executing==1.2.0
fastapi==0.116.2
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastjsonschema==2.16.3
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
google-auth==2.42.1
google-genai==1.47.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.4
importlib-metadata==6.1.0
importlib_resources==6.4.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.27
jaxlib==0.4.27
jedi==0.18.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.0.3
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.20.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.13
mistune==2.0.5
ml-dtypes==0.4.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
orjson==3.11.3
packaging==23.0
pandas==2.3.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil==5.9.4
psycopg2-binary==2.9.11
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.14.0
pyparsing==3.1.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.1.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2025.2
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.1
qtconsole==5.4.1
QtPy==2.3.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.0
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.4
SQLAlchemy==2.0.44
stack-data==0.6.2
starlette==0.48.0
tenacity==9.1.2
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.5
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.15.0
